record refactor
